新增试用功能只读JSON API：抽取trial页计算逻辑为共享服务，老页面HTML逐字节不变
