+++ app/api/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 api_bp = Blueprint('api', __name__)
 
-from . import stock_api, analysis_api, text2sql_api 
+from . import stock_api, analysis_api, text2sql_api, trial_api
